Move focus switcher to a separate file
